backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0